pyproject.toml
[project]
name = "ai-cax-110-capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
# System dependency: ffmpeg (install via: brew install ffmpeg)
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.34.0",
    "groq>=1.1.1",
    "langdetect>=1.0.9",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
]
